+++ src/components/pages/home/documents-section.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { IoDocuments } from 'react-icons/io5';
 
-import { Button } from '@/components/shared/Button';
-import HeadingH1 from '@/components/shared/Heading';
-import List from '@/components/shared/List';
+import { Button } from '@/components/shared/button';
+import HeadingH1 from '@/components/shared/heading';
+import List from '@/components/shared/list';
 import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area';
 import { documentsData } from '@/utils/static-resources/documents.static';
