+++ src/components/pages/home/executives-section.tsx
@@ -1,12 +1,12 @@
 import { useMemo, useState } from 'react';
 import { PiArrowCircleUpRightFill } from 'react-icons/pi';
 
-import { Button } from '@/components/shared/Button';
-import { CardWrapper } from '@/components/shared/CardWrapper';
-import { ExecutivesCard } from '@/components/shared/ExecutivesCard';
-import HeadingH1 from '@/components/shared/Heading';
+import { Button } from '@/components/shared/button';
+import { CardWrapper } from '@/components/shared/card-wrapper';
+import { ExecutivesCard } from '@/components/shared/executives-card';
+import HeadingH1 from '@/components/shared/heading';
 
-import CustomPagination from '@/components/shared/Pagination';
+import CustomPagination from '@/components/shared/pagination';
 import { executivesData } from '@/utils/static-resources/executives.static';
 
 const Executives = () => {
