+++ src/components/pages/home/management-structureSection.tsx
@@ -1,7 +1,7 @@
 import Document from '/assets/images/Document.png';
 
-import HeadingH1 from '@/components/shared/Heading';
-import List from '@/components/shared/List';
+import HeadingH1 from '@/components/shared/heading';
+import List from '@/components/shared/list';
 
 const ManagementStructure = () => {
   return (
