+++ src/components/pages/home/partners-section.tsx
