+++ src/components/shared/Suspense.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Loader from './Loader';
+import Loader from './loader';
 
 interface SuspenseProps {
   children: React.ReactNode;
