+++ src/components/shared/executives-card.tsx
