+++ src/components/ui/sonner.tsx
@@ -5,25 +5,22 @@ import { Toaster as Sonner } from 'sonner';
 type ToasterProps = React.ComponentProps<typeof Sonner>;
 
 const Toaster = ({ ...props }: ToasterProps) => {
-  const { theme = 'system' } = useTheme();
+  const { theme } = useTheme();
 
   return (
     <Sonner
-      richColors
       position={props.position || 'top-right'}
-      theme={theme as ToasterProps['theme']}
+      richColors
+      theme={theme === 'dark' ? 'dark' : 'light'}
       className="toaster group"
+      duration={props.duration || 3000}
       toastOptions={{
-        classNames: {
-          toast:
-            'group toast group-[.toaster]:bg-background group-[.toaster]:text-foreground group-[.toaster]:border-border group-[.toaster]:shadow-lg',
-          description: 'group-[.toast]:text-muted-foreground',
-          actionButton:
-            'group-[.toast]:bg-primary group-[.toast]:text-primary-foreground',
-          cancelButton:
-            'group-[.toast]:bg-muted group-[.toast]:text-muted-foreground',
+        className: 'font-montserrat  ',
+        style: {
+          padding: '12px',
         },
       }}
+      closeButton
       {...props}
     />
   );
